move everything related to the old parser to archive and update references
