+++ steel/src/new_parser/mod.rs
@@ -5,6 +5,7 @@ pub mod lexer;
 pub mod parser;
 pub mod rename_idents;
 pub mod replace_idents;
+pub mod span;
 pub mod tokens;
 pub mod tryfrom_visitor;
 pub mod visitors;
