+++ steel_vm/src/inline_iter.rs
@@ -6,7 +6,7 @@ use std::rc::Rc;
 use steel::core::instructions::DenseInstruction;
 use steel::env::Env;
 use steel::gc::Gc;
-use steel::parser::span::Span;
+use steel::new_parser::span::Span;
 use steel::rerrs::SteelErr;
 use steel::rvals::{Result, SteelVal};
 use steel::steel_compiler::constants::ConstantTable;
