+++ steel_vm/src/vm.rs
@@ -29,10 +29,13 @@ use steel::primitives::{ListOperations, VectorOperations};
 use steel::rerrs::SteelErr;
 use steel::rvals::{ByteCodeLambda, Result, SteelVal};
 // use steel_compiler::expand::MacroSet;
-// use steel::parser::span::Span;
-use steel::parser::{Expr, ParseError, Parser};
+// use steel::new_parser::span::Span;
+// use steel::parser::{Expr, ParseError, Parser};
 
 use steel::structs::SteelStruct;
+
+use steel::new_parser::ast::ExprKind;
+use steel::new_parser::parser::{ParseError, Parser};
 
 // use std::cell::Cell;
 // use steel::env::CoreModuleConfig;
@@ -497,7 +500,7 @@ pub fn vm<CT: ConstantTable>(
                     // TODO
                     let mut intern = HashMap::new();
 
-                    let parsed: result::Result<Vec<Expr>, ParseError> =
+                    let parsed: result::Result<Vec<ExprKind>, ParseError> =
                         Parser::new(expr.as_str(), &mut intern).collect();
 
                     match parsed {
